# api/openapi.yaml
openapi: 3.1.0
info:
  title: Travel-Planning Agent — Frontend API
  version: 1.0.0
  license:
    name: UNLICENSED
  description: |
    HTTP + SSE contract between the Vue 3 chat frontend and the travel-planning agent backend.

    The agent picks optimal flights / hotels / tours for a travel group given budget, constraints
    and preferences, asks closed clarifying questions when needed, renders visited settlements as
    map points, and rebuilds the plan when the user edits the route.

    **Design reference:** `docs/superpowers/specs/2026-06-13-travel-agent-api-design.md`
    **Module spec:** `api/SPECIFICATION.md`

    ## Conventions
    - **Auth:** HTTP `Bearer` JWT (access token) on every endpoint except
      `register` / `login` / `refresh` and the SSE stream (which uses a one-time `ticket`).
    - **Real-time:** a chat message creates an async *run*; the run's events are delivered over
      Server-Sent Events (`text/event-stream`). Reconnect with the `Last-Event-ID` header.
    - **Money:** integer `*_rub` fields are whole rubles, matching the dataset.
    - **IDs:** opaque strings (UUIDs for new entities; seeded groups keep `G-0001`-style ids).
    - **Timestamps:** RFC 3339 UTC, except flight `departure_time`/`arrival_time` which are local
      `HH:MM` time-of-day strings (as stored in `flights.csv`).
    - **Errors:** every non-2xx response uses the `Error` envelope.
...
      backing CSVs store them as `0`/`1`.

    ## Out of scope (v1)
...
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /chat:
    post:
      tags: [Chat]
      summary: Send a message or answer a clarifying question (creates a run)
      operationId: postChat
      description: |
        Starts a new agent turn (*run*). Omit `session_id` to start a new chat.

        Provide **either** `message`, **or** an answer to a clarifying question
        (`in_reply_to_question_id` plus `selected_option_ids` and/or `freeform`).

        The response returns immediately with a `run_id`; subscribe to results by minting a
        stream ticket (`POST /chat/{run_id}/stream-ticket`) and opening the SSE stream.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChatRequest' }
      responses:
        '202':
          description: Run accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChatAccepted' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /chat/{run_id}/stream-ticket:
    post:
      tags: [Chat]
      summary: Mint a short-lived, single-use ticket for the SSE stream
      operationId: createStreamTicket
      description: |
        Browser `EventSource` cannot set an `Authorization` header, so the SSE stream is authorized
        by a short-lived, single-use, run-scoped ticket minted here with the normal Bearer token.
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Stream ticket
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StreamTicket' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /chat/{run_id}/stream:
    get:
      tags: [Chat]
      summary: Stream run events (Server-Sent Events)
      operationId: streamRun
      security:
        - streamTicket: []
      description: |
        Opens a `text/event-stream`. Authorized by the `ticket` query parameter (see
        `stream-ticket`). Each SSE frame has an `event:` name and a JSON `data:` payload, and
        carries an `id:` so a dropped connection can resume via the `Last-Event-ID` request header.
        The stream closes when a terminal `run_status` (`completed` / `cancelled` / `error`) is sent.

        **Event names → data schema:**
        | `event:` | `data:` schema |
        |---|---|
        | `run_status` | `RunStatusEvent` |
        | `message_delta` | `MessageDeltaEvent` |
        | `message` | `MessageEvent` |
        | `clarifying_question` | `ClarifyingQuestionEvent` |
        | `plan_status` | `PlanStatusEvent` |
        | `map` | `MapEvent` |
        | `error` | `ErrorEvent` |

        While `plan_status` is not `ready`, the frontend keeps the map read-only.
      parameters:
        - $ref: '#/components/parameters/RunId'
        - name: ticket
          in: query
          required: true
          description: One-time stream ticket from `POST /chat/{run_id}/stream-ticket`.
          schema: { type: string }
        - name: Last-Event-ID
          in: header
          required: false
          description: Resume the stream after this event id (set automatically by `EventSource`).
          schema: { type: string }
      responses:
        '200':
          description: |
            An event stream. The schema below documents the union of possible `data:` payloads;
            use the `event:` name to discriminate (see the table above).
          content:
            text/event-stream:
              schema: { $ref: '#/components/schemas/SseEvent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /chat/{run_id}/cancel:
    post:
      tags: [Chat]
      summary: Cancel an in-flight run
      operationId: cancelRun
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '202':
          description: Cancellation requested
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CancelResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Run already finished
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ───────────────────────────── Sessions ─────────────────────────────
...
  /plans/{plan_id}/modify:
    post:
      tags: [Plans]
      summary: Correct the route — add/remove map points as a batch and rebuild
      operationId: modifyPlan
      description: |
        Allowed only while the plan is `ready`. Submits the user's batch of added/removed map points
        (plus an optional note); the plan flips back to `building` and the agent rebuilds. Subscribe
        to the **new** run's stream (`run_id` in the response) for the rebuild events.
      parameters:
        - $ref: '#/components/parameters/PlanId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ModifyRequest' }
      responses:
        '202':
          description: Rebuild started on a new run
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ModifyAccepted' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/PlanNotReady' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token (`sub = user_id`, ~15 min lifetime).
    streamTicket:
      type: apiKey
      in: query
      name: ticket
      description: Single-use, run-scoped ticket for the SSE stream (~60 s lifetime).
  parameters:
...
    RunId:
      name: run_id
      in: path
      required: true
      schema: { type: string }
...
    PlanId:
      name: plan_id
      in: path
      required: true
      schema: { type: string }
  responses:
    Unauthorized:
      description: Missing/invalid token, or token expired
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but not allowed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Unknown resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    PlanNotReady:
      description: Modify/accept attempted while the plan is not `ready`
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Malformed or invalid body/params
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - validation_error
                - unauthorized
                - token_expired
                - forbidden
                - not_found
                - conflict
                - plan_not_ready
                - rate_limited
                - internal
            message: { type: string, description: Human-readable explanation }
            details:
              type: object
              additionalProperties: true
              description: Optional machine-readable detail (e.g. per-field errors).

    # ── Auth ──
...
    ChatRequest:
      type: object
      description: |
        Provide `message`, OR answer a clarifying question with `in_reply_to_question_id` plus
        `selected_option_ids` and/or `freeform`.
      properties:
        session_id:
          type: string
          description: Omit to start a new chat (session).
        message: { type: string, description: User text }
        group_id:
          type: string
          description: Selected travel-group context; attached to the resulting plan.
        in_reply_to_question_id:
          type: string
          description: Set when answering a clarifying question.
        selected_option_ids:
          type: array
          items: { type: string }
          description: Chosen closed-answer option id(s).
        freeform:
          type: string
          description: Freeform answer when the offered options don't fit.
    ChatAccepted:
      type: object
      required: [run_id, session_id]
      properties:
        run_id: { type: string }
        session_id: { type: string }
    StreamTicket:
      type: object
      required: [ticket, expires_in]
      properties:
        ticket: { type: string }
        expires_in: { type: integer, description: Ticket lifetime in seconds, example: 60 }
    CancelResponse:
      type: object
      required: [run_id, status]
      properties:
        run_id: { type: string }
        status: { type: string, const: cancelling }

    # ── SSE events ──
    RunStatus:
      type: string
      enum: [started, running, completed, cancelled, error]
    PlanBuildStatus:
      type: string
      enum: [building, ready, error]
    MapPoint:
      type: object
      required: [id, name, kind, lat, lng, order]
      description: |
        Ordered route point. Rich place-card fields are optional Agent Service content validated
        and persisted by the backend. All strings are plain text, never trusted HTML.
      properties:
        id: { type: string }
        name: { type: string, description: Place / route point label }
        kind: { type: string, enum: [origin, destination, stop] }
        lat: { type: number, format: double }
        lng: { type: number, format: double }
        order: { type: integer, description: Position along the route (0-based) }
        note: { type: string, maxLength: 2000 }
        description:
          type: string
          maxLength: 1000
          deprecated: true
          description: Legacy alias of `summary`; both carry the same value when present.
        summary: { type: string, maxLength: 1000 }
        visit_date: { type: string, format: date }
        visit_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
          description: Local visit time in HH:MM form.
        visit_start: { type: string, format: date-time }
        visit_end: { type: string, format: date-time }
        duration_minutes: { type: integer, minimum: 0, maximum: 10080 }
        cost_rub: { type: integer, minimum: 0 }
        price_note: { type: string, maxLength: 1000 }
        calendar_event_id:
          type: string
          description: Backend-owned id of the linked calendar event.
        ref_id:
          type: string
          description: Related public business reference, when applicable.
        transport_to_next: { type: string, maxLength: 1000 }
        travel_time_to_next_minutes: { type: integer, minimum: 0, maximum: 10080 }
        distance_to_next_km: { type: number, format: double, minimum: 0, maximum: 50000 }
        historical_background: { type: string, maxLength: 4000 }
        interesting_facts:
          type: array
          maxItems: 20
          items: { type: string, minLength: 1, maxLength: 500 }
        visit_tips:
          type: array
          maxItems: 20
          items: { type: string, minLength: 1, maxLength: 500 }
        food_recommendations:
          type: array
          maxItems: 20
          items: { type: string, minLength: 1, maxLength: 500 }
        signature_dishes:
          type: array
          maxItems: 20
          items: { type: string, minLength: 1, maxLength: 500 }
        average_check_rub: { type: integer, minimum: 0 }
        booking_advice: { type: string, maxLength: 2000 }
        accessibility_notes: { type: string, maxLength: 2000 }
        safety_notes: { type: string, maxLength: 2000 }
        weather_notes: { type: string, maxLength: 2000 }
        why_recommended: { type: string, maxLength: 2000 }
        content_source:
          type: string
          maxLength: 100
          pattern: '^[a-z][a-z0-9_:-]*$'
        content_confidence:
          type: string
          enum: [low, medium, high]
    QuestionOption:
      type: object
      required: [id, label]
      properties:
        id: { type: string }
        label: { type: string }
    ClarifyingQuestion:
      type: object
      required: [id, text, options, allow_freeform]
      properties:
        id: { type: string }
        text: { type: string }
        options:
          type: array
          description: Mutually exclusive closed-answer options.
          items: { $ref: '#/components/schemas/QuestionOption' }
        allow_freeform:
          type: boolean
          description: Whether a freeform answer is accepted when no option fits.
    RunStatusEvent:
      type: object
      required: [run_id, status]
      properties:
        run_id: { type: string }
        status: { $ref: '#/components/schemas/RunStatus' }
    MessageDeltaEvent:
      type: object
      required: [run_id, message_id, delta]
      properties:
        run_id: { type: string }
        message_id: { type: string }
        delta: { type: string, description: Streamed assistant text chunk }
    MessageEvent:
      type: object
      required: [run_id, message]
      properties:
        run_id: { type: string }
        message: { $ref: '#/components/schemas/Message' }
    ClarifyingQuestionEvent:
      type: object
      required: [run_id, question]
      properties:
        run_id: { type: string }
        question: { $ref: '#/components/schemas/ClarifyingQuestion' }
    PlanStatusEvent:
      type: object
      required: [run_id, plan_id, status]
      properties:
        run_id: { type: string }
        plan_id: { type: string }
        status: { $ref: '#/components/schemas/PlanBuildStatus' }
        error: { type: string, description: Present when status is error }
    MapEvent:
      type: object
      required: [run_id, plan_id, points]
      description: Full map snapshot (not a delta).
      properties:
        run_id: { type: string }
        plan_id: { type: string }
        points:
          type: array
          items: { $ref: '#/components/schemas/MapPoint' }
    ErrorEvent:
      type: object
      required: [run_id, error]
      properties:
        run_id: { type: string }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
    SseEvent:
      description: |
        Union of possible SSE `data:` payloads. Discriminate on the SSE `event:` name (see the
        stream operation's description), not on a body field.
      oneOf:
        - $ref: '#/components/schemas/RunStatusEvent'
        - $ref: '#/components/schemas/MessageDeltaEvent'
        - $ref: '#/components/schemas/MessageEvent'
        - $ref: '#/components/schemas/ClarifyingQuestionEvent'
        - $ref: '#/components/schemas/PlanStatusEvent'
        - $ref: '#/components/schemas/MapEvent'
        - $ref: '#/components/schemas/ErrorEvent'

    # ── Sessions ──
    Answer:
      type: object
      required: [in_reply_to_question_id]
      description: A user reply to a clarifying question.
      properties:
        in_reply_to_question_id: { type: string }
        selected_option_ids:
          type: array
          items: { type: string }
        selected_option_labels:
          type: array
          items: { type: string }
          readOnly: true
          description: Labels resolved by the backend for persisted answer history.
        freeform: { type: string }
    PlanRef:
      type: object
      required: [plan_id, status]
      properties:
        plan_id: { type: string }
        status: { $ref: '#/components/schemas/PlanStatus' }
    Message:
      type: object
      required: [id, role, content, created_at]
      properties:
        id: { type: string }
        role: { type: string, enum: [user, assistant] }
        content: { type: string }
        created_at: { type: string, format: date-time }
        run_id: { type: string }
        question:
          allOf: [{ $ref: '#/components/schemas/ClarifyingQuestion' }]
          description: Present on an assistant message that asks a clarifying question.
        answer:
          allOf: [{ $ref: '#/components/schemas/Answer' }]
          description: Present on a user message that answers a clarifying question.
        plan_ref:
          allOf: [{ $ref: '#/components/schemas/PlanRef' }]
          description: Present on an assistant message that (re)built a plan.
...
    PlanStatus:
      type: string
      enum: [building, ready, error, accepted, rejected]
...
    AddPoint:
      type: object
      required: [name]
      properties:
        name: { type: string }
        kind: { type: string, enum: [origin, destination, stop] }
        lat: { type: number, format: double }
        lng: { type: number, format: double }
        after_point_id:
          type: string
          description: Insert after this existing point (route ordering hint).
    ModifyRequest:
      type: object
      description: Batch route correction. The plan must be `ready`.
      properties:
        add:
          type: array
          items: { $ref: '#/components/schemas/AddPoint' }
        remove:
          type: array
          description: Ids of map points to remove.
          items: { type: string }
        note:
          type: string
          description: Free-text wishes for the rebuild.
    ModifyAccepted:
      type: object
      required: [run_id]
      properties:
        run_id: { type: string, description: The rebuild run to stream. }
...